Fix Unix test failures
